feat: add runtime config, improve Docker image, upgrade frontend
This commit adds the runtime configuration instead of embedding
a `.env` file and `yarn build` every time the container starts.

With this change, `envsubst` is used, a `config.json` file is
created using the `config.template.json` file that will then
be populated at runtime before starting the nginx server.

Due to this change, the client has to perform an HTTP GET request
to the server to get the `config.json` file before loading the
Vue application.

Additionally, this commit upgrades some yarn dependencies
and sets the default map style to one of the MapBox's defaults
(the author's one is not working for some reason).

fix: revert zoom hardcoded change